pyproject.toml
[project]
name = "DikeBenchmarker"
version = "0.1.0"
description = "Sustainable Competition"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.8.0",
    "py-cpuinfo>=9.0.0",
    "scikit-learn>=1.4.1.post1",
    "parsl>=2025.9.1",
    "gbd-tools>=5.0.0",
    "pyyaml",
]

[tool.ruff]
line-length = 160
select = ["E", "F", "D"]  # enable docstring rules

[tool.ruff.pydocstyle]
convention = "google"

[tool.pylint.format]
max-line-length = 160
max-args = 11
max-positional-arguments = 11

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.12.11",
    "requests>=2.31.0",
    "pytest-dependency<0.6.0",
]

[tool.setuptools.packages.find]
where = ["src"] # This tells setuptools to look in src/

[tool.setuptools.package-data]
"DikeBenchmarker" = ["DikeBenchmarker/data/db/sustainablecompetition.db"]
